Cleaned up the code for readability.
Currently have a working function to display all tweets on a button click,
and a function to display a single user's tweets (hardcoded, needs to be
updated to pass in a variable) when a username is clicked on.

Have written what I think would be a working function (tweetStream) if I can
figure out how to grab and pass a username to it.